Feature/fargate web app (#2)
* Testing build docker image in CICD

* testing docker build and push to ECR

* Introduced IAM role for logging in / pushing to ECR

* added role to codebuild step

* changed role principal to codebuild

* Passing CF outputs as environment variables to web stack

* Upload endpoint and key to SSM

* Added ssm:GetParameter to CodeBuild permissions

* removed sed command

* Create .env.local in CodeBuild project

* Testing ECS LoadBalancedFargate pattern

* Fixed eslint errors in web app

* Refactored web stack, introduced networking stack

* Remove throw if API endpoint / key are undefined

* Introduced empty list component

* Retrieve env vars from SSM at build time

* Asserting endpoint details are defined in server actions

* Improved error logging and replaced SSM parameters for CFN outputs

* Updated endpoint variable to use correct API resource, improved logging.

* Removed API key from API GW resource

* testing cache revalidation

* CodeDeploy B/G ECS deployment test

* Introduced scripts to prepare file set for B/G CodeDeploy step

* Updated codebuild directory

* Cleaned up code and fixed B/G deployment

* Reintroduced task def execution role

* Updated code deploy configuration script

* ecs all at once test

* Using canary deployment config

* Updated configuration script to pick up the correct IAM role

* Updated config step to grab task execution role

* ECS logging test

* Using the correct log group permission name

* Adding endpoint to generated task def file

* Updated CodeDeploy wait times

* Moved shell scripts into separate directory

* Refactored BuildAndUploadDockerImage as a ShellStep

* Refactored back to CodeBuildStep

* Cleaned up stateless, pipeline, and web stacks

* Replaced endpoint CfnOuput with SSM param

* Gave configure b/g step SSM permissions

* More clean up and comments

* Cleaned up scripts directory

* Fixed task def template JSON

* Fixed incorrect SSM ARN

* Remove double quotes from SSM value

* README update and switch pipeline to point to main

+++ web/components/TodoSection/index.tsx
@@ -13,6 +13,21 @@ import {
 } from "@/app/actions";
 import { ToastQueueContext } from "../ToastQueueProvider";
 
+function EmptySection() {
+  return (
+    <div className={utilStyles.centredText}>
+      <h2 className={utilStyles.headingXl}>
+        🎉 Congrats, your Todo list is{" "}
+        <span className={utilStyles.highlightedText}>empty!</span>
+      </h2>
+      <p>
+        <span className={utilStyles.italicText}>Forget something?</span> Add a
+        new entry to your list above.
+      </p>
+    </div>
+  );
+}
+
 type TodoSectionProps = {
   entries: TodoEntry[];
 };
@@ -22,6 +37,7 @@ export default function TodoSection(props: TodoSectionProps) {
 
   const { enqueueToast } = useContext(ToastQueueContext);
 
+  // Below are our 3 server actions for creating, updating, and deleting entries respectively.
   function addEntry(formData: FormData) {
     serverPostEntry(formData)
       .then((newEntry) => {
@@ -31,8 +47,13 @@ export default function TodoSection(props: TodoSectionProps) {
           message: "Successfully created new todo entry.",
         });
       })
-      .catch((error) => {
-        enqueueToast({ level: "error", message: `${error}`, lifespan: "inf" });
+      .catch(() => {
+        enqueueToast({
+          level: "error",
+          message:
+            "Sorry, we had trouble creating your Todo entry, please try again later.",
+          lifespan: "inf",
+        });
       });
   }
 
@@ -51,8 +72,13 @@ export default function TodoSection(props: TodoSectionProps) {
           message: "Successfully updated todo entry.",
         });
       })
-      .catch((error) => {
-        enqueueToast({ level: "error", message: `${error}`, lifespan: "inf" });
+      .catch(() => {
+        enqueueToast({
+          level: "error",
+          message:
+            "Sorry, we had trouble updating your Todo entry, please try again later.",
+          lifespan: "inf",
+        });
       });
   }
 
@@ -67,35 +93,50 @@ export default function TodoSection(props: TodoSectionProps) {
           message: "Successfully deleted todo entry.",
         });
       })
-      .catch((error) => {
-        enqueueToast({ level: "error", message: `${error}`, lifespan: "inf" });
+      .catch(() => {
+        enqueueToast({
+          level: "error",
+          message:
+            "Sorry, we had trouble deleting your Todo entry, please try again later.",
+          lifespan: "inf",
+        });
       });
+  }
+
+  let body = <EmptySection />;
+
+  if (entries.length) {
+    body = (
+      <>
+        <div className={styles.headingBar}>
+          <div className={styles.firstHeading}>
+            <h2 className={utilStyles.headingLg}>Description</h2>
+          </div>
+          <div className={styles.secondHeading}>
+            <h2 className={utilStyles.headingLg}>Date Added</h2>
+          </div>
+          <div className={styles.fillerBlock}></div>
+        </div>
+        <div>
+          {entries.map((entry, i) => (
+            <Fragment key={entry.Id}>
+              <TodoItem
+                item={entry}
+                checkItem={() => updateEntry(entry.Id, !entry.Completed)}
+                deleteItem={() => removeEntry(entry.Id)}
+              />
+              {i < entries.length - 1 && <hr />}
+            </Fragment>
+          ))}
+        </div>
+      </>
+    );
   }
 
   return (
     <div className={styles.table}>
       <AddTodoEntryForm formAction={addEntry} />
-      <div className={styles.headingBar}>
-        <div className={styles.firstHeading}>
-          <h2 className={utilStyles.headingLg}>Description</h2>
-        </div>
-        <div className={styles.secondHeading}>
-          <h2 className={utilStyles.headingLg}>Date Added</h2>
-        </div>
-        <div className={styles.fillerBlock}></div>
-      </div>
-      <div>
-        {entries.map((entry, i) => (
-          <Fragment key={entry.Id}>
-            <TodoItem
-              item={entry}
-              checkItem={() => updateEntry(entry.Id, !entry.Completed)}
-              deleteItem={() => removeEntry(entry.Id)}
-            />
-            {i < entries.length - 1 && <hr />}
-          </Fragment>
-        ))}
-      </div>
+      {body}
     </div>
   );
 }
